feat: change behaviour of fetch
fix #372

this PR will add a new method in the `context`. THe new method `gracefullyFetch` is a replace of the previous method `fetch`.

`fetch` method now will return the `response` without any transformation.
`gracefullyFetch` will return `json` or `text` depending on the header

This will be a breaking change for ones using the `fetch` method. I thought to use a parameter instead of creating a new method but it makes things less clear.